+++ Backend/Zetta.Infraestructure/Repositories/GenericRepository/IGenericRepository.cs
@@ -8,5 +8,5 @@ public interface IGenericRepository<T> where T : class
     Task<T> GetFilterFirstOrDefaultAsync(Expression<Func<T, bool>> predicate);
     Task<T> AddAsync(T entity);
     Task<T> UpdateAsync(T entity);
-    Task<T> DeleteAsync(long id);
+    Task<T> DeleteAsync(int id);
 }
