+++ Backend/Zetta.Infraestructure/Repositories/TimeValueRepository/ITimeValueRepository.cs
@@ -7,5 +7,5 @@ public interface ITimeValueRepository
     Task<IEnumerable<TimeValue>> GetTimeValuesAsync();
     Task<TimeValue> AddTimeValueAsync(TimeValue timeValue);
     Task<TimeValue> UpdateTimeValueAsync(TimeValue timeValue);
-    Task<TimeValue> DeleteTimeValueAsync(long id);
+    Task<TimeValue> DeleteTimeValueAsync(int id);
 }
